refactor: remove config in file names
